services/receipt-assembler/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.5
google-cloud-storage==3.2.0
google-auth==2.48.0
